(PC-25607) refactor(offer): rename offerPlaylist to offerPlaylistOld for offerV1
