+++ src/features/offer/components/OfferPlaylistOld/api/fetchOffersByArtist.test.ts
@@ -4,7 +4,7 @@ import { SearchGroupNameEnumv2 } from 'api/gen'
 import {
   buildAlgoliaFilter,
   fetchOffersByArtist,
-} from 'features/offer/components/OfferPlaylist/api/fetchOffersByArtist'
+} from 'features/offer/components/OfferPlaylistOld/api/fetchOffersByArtist'
 import { offerAttributesToRetrieve } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/offerAttributesToRetrieve'
 
 jest.mock('algoliasearch')
